backend/requirements.txt
langchain==0.1.6
transformers==4.37.2
fastapi==0.109.2
python-dotenv==1.0.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
unstructured==0.12.4
together==0.2.11
flashrank
uvicorn==0.27.1
pymongo==4.6.3
bcrypt==4.1.2
torch==2.2.2
tensorboard==2.16.2
pandas==2.2.1
matplotlib==3.8.4
python-dotenv==1.0.1
google-generativeai==0.4.1
python-multipart==0.0.9